+++ src/sdk/dto/get-transactions.dto.ts
@@ -1,0 +1,11 @@
+import { IsPositive } from 'class-validator';
+import { IsAddress } from './validators';
+import { PaginationDto } from './pagination.dto';
+
+export class GetTransactionsDto extends PaginationDto {
+  @IsAddress()
+  account: string = null;
+
+  @IsPositive()
+  chainId: number;
+}