+++ src/sdk/dto/index.ts
@@ -13,3 +13,4 @@ export * from './get-token-list.dto';
 export * from './pagination.dto';
 export * from './get-exchange-supported-assets.dto';
 export * from './get-token-lists.dto';
+export * from './get-transactions.dto';