+++ CHANGELOG.md
@@ -1,4 +1,10 @@
 # Changelog
+## [1.6.4] - 2024-03-20
+### New
+- Added `getTransactions` endpoint into DataUtils
+### Breaking changes
+- The parameters named `projectKey` and `graphqlEndpoint` have been removed in the PrimeSdk module
+
 ## [1.6.3] - 2024-03-19
 ### New
 - Added ArkaPaymaster as a sub-module